% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADRCDashPublic-package.R
\docType{package}
\name{ADRCDashPublic-package}
\alias{ADRCDashPublic}
\alias{ADRCDashPublic-package}
\title{ADRCDashPublic: ADRC Public Facing Dashboard}
\description{
Data explorer dashboard for the University of Alabama at Birmingham Alzheimer's Disease Center.
}
\author{
\strong{Maintainer}: Chad Murchison \email{AUTHOR@MAIL.COM}

Authors:
\itemize{
  \item Joseph Marlo \email{support@landeranalytics.com}
}

}
\keyword{internal}
